Refactor vm allocation log
The IPv4 allocation follows this line. Even if we print this, we may not
allocate a VM on this host. It's best to print when the allocation is
finalized.

In the past, we encountered concurrency and lease violation problems
with respirate, resulting in the same IP being assigned to multiple VMs.
We included detailed information in the log to assist with debugging.
However, these values have become too verbose and are no longer useful.