Fix veth pair setup idempotency bug
With this commit, we are fixing the idempotency bug we encountered at
the network interfaces setup. We relied on namespace delete to remove
the veth pairs before attempting to recreate them in case of a failure.
However, looks like this is not always true. Especially when there is an
active process, running in the namespace, the delete command returns
success but the interface is not removed until the process is cleaned.
That's why we get the failure when trying to add the veth interface with
the same name. Here, we had 2 choices, either to remove the interface
manually, or kill the processes inside the namespace before trying to
delete it. We chose the latter one to not leak processes.