Fix veth pair setup idempotency bug
With this commit, we are fixing the idempotency bug we encountered at
the network interfaces setup. We relied on namespace delete to remove
the veth pairs before attempting to recreate them in case of a failure.
However, looks like this is not always true. The main problem is, the
network namespace deletion does not handle the related interfaces
deletion in a synchronous way. The command is to delete the namespace is
not blocking. Therefore, the immediate interface creation may fail if
executed quickly. To avoid this problem, we are introducing a loop to
check the existence of the interface up to 5 times.