Initialize rodauth per documentation
Normally, you are to call `#rodauth` first, before using its features
such as `load_memory`.  Now it seems that the way we had it worked as
expected, but it was contra-documentation.

To double check, checked the Rodauth test suite, which has code like
this in `password_grace_period_spec.rb`, so indeed, we were unusual.

    roda do |r|
      r.rodauth
      rodauth.load_memory
      r.root do