Make number of SPDK cpus a function of host cpu count.
Previously we always used 2 cpus for SPDK. But this is not ideal since
in larger hosts we can fit more VMs, which meant less I/O capacity. To
address this, we make SPDK cpu count a function of host cpu count and
allocate 4 SPDK cpus when host cpus are more than 64, and 2 otherwise.

This means we'll have 2 SPDK cpus in AX161 which has 64 cpus, but 4 SPDK
cpus in AX162 which has 96 cpus. Number of SPDK cpus will also be 4 in
RX220, which has 80 host cpus.

Also note that this doesn't further decrease the VM capacity in AX162,
since in practice we could reserve 374 out of 384 hugepages which
already reduced VM capacity to 46.