Avoid VM allocation and unnecessary VM reboots when host rebooted
When we start rebooting a host, we should make sure that no new VM
allocations are happenning in this vm host. This causes increased number
of vm allocations to have higher vm provisioning time.

Furthermore, if a VM is not even prepped yet, there is no reason to
reboot it after the host is rebooted, because it will only get properly
allocated only after the host reboots. If we don't do this, we will try
to recreate interfaces, hugepages and storage for an already up and
running VM. This will eventually cause idempotency issues.

This does not solve the issue at hand completely but minimize the amount
of VMs impacted after a host reboot. The ideal solution is that if a VM
did not get provisioned fully, at the time of host reboot, we should
reallocate it on another host after performing proper clean-up. Though,
this would require proper record and host clean-up and the storage
information to be rewritten into the frame, since we clean that up after
the allocation.