Optimize thread usage in monitor
In monitor, we create 2 threads per resource; one for SSH processing and one
for pulse check. In previous version, each resource would keep their threads
even after the pulse check is completed. This means in an environment where
there is a limit on number of threads, we can monitor maximum of limit / 2
resources at the same time. This commit changes the behavior so that after the
pulse check is completed, the threads are terminated. This way, we can monitor
significantly more resources at the same time.
One drawback of the new approach is that we need to re-create the threads for
each check. In my system creating 1000 threads takes about 0.025 seconds, so
overhead is seems negligible.
Another change we are making is that removing the monitoring_interval from the
model and hardcoding it in the monitor as 5 seconds. This removes capability of
setting different monitoring intervals for different resources. Supporting this
would require some work and since it is not used in the current implementation
I decided to remove all together. If we need to support this in the future, we
can add it back with some effort.