Fix GithubRunner setup_environment actions-runner script idempotency
In case the script is run until we mv to /home/runner but the state
could not progress because of any reason (e.g. deployment, runner
failure), the idempotency is broken. The script needs to be copied back
to the home directory of the runneradmin. More information regarding the
operation can be found in the middle of the script where we chown the
actions-runner.