Monitor Vm resources
This just monitors the Vm resources, and creates page when there is a failure.
It does not try to heal it. After using this for a while in prod, we can decide
on how the automated healing would look like.
For monitoring purposes, we only focus on Vms that does not have an associated
Sshable entity. This is because Vms with Sshable entity are children of higher
level resources like PostgresServer or GithubRunner and will be monitored by
the their respective parent resource. This is a safeguard against duplicate
page creation for the same problem. When we are much bigger organization where
there are specialized teams for each resource, it would make sense to create
separate pages for Vms and their parent resources, because each team would
independently try to heal their resources. However, for now, this is not
necessary.