Nap randomly when concurrency limit reached
When customer concurrency limit is reached, we run a utilization check.
If the utilization is low enough, we allow them to provision the VM,
anyway. However, if the utilization is higher than a set value, we used
to nap 5 seconds and run the same cycle. The problem with this approach
is that when we get a flood of requests in the same time, generally, the
utilization check times are almost the same. However, the utilization
value only changes when the VM entity is created and allocated.
Therefore, we cannot really catch the real utilization. To be more
clear, let's say we had a concurrency limit of 100 vcpus and since the
global utilization was low, we could get up to 150 vcpu usage. On top of
it, we have 10 more runners waiting to get concurrency limit free up or
the utilization to get below 70%. The moment 1 VM gets deprovisioned and
the utilization gets to (let's say) 69%, those 10 runners almost
simultenaously check the utilization and get the permission to provision
the VMs. This creates another flooding in the system. We do not want to
get strict locks on VmHosts to learn the utilization, hence, I think
sleeping the processes with random intervals will help reduce the
impact.