Do not try to delete runner if script already running
We catch Octokit::Conflict exception with the "Already exists" message
to ensure the `register_runner` function is idempotent. If it's run a
second time, it fails with this exception.

Recently, we found a few examples with non-idempotent results. If this
label run again after the start of the script, it fails with the
following error:

    Failed to start transient service unit: Unit runner-script.service
    already exists

This PR fixes this issue by checking if the script is already running
before trying to delete the runner.