Wait until no virtual machines with the given image while downloading new image
Since we provision virtual machines using copy-on-write, we avoid
deleting the base image when any virtual machines are booted from it. We
wait until there are no virtual machines on the host before downloading
new boot images. This approach works for the runner hosts. However,
after fleet unification, if a regular VM or PostgreSQL database is on
the host, full draining isn't possible. So, we can wait until no virtual
machines are using the given boot image on the host. Virtual machines
from other images shouldn't be affected.