Allow to generate invoices for specific projects
We use InvoiceGenerator to display the project's current usage to the
customer. We use the 'project_id' argument to filter the project. If
it's not provided, we generate invoices for all projects.

Occasionally, we need to generate invoices for specific projects. To
accommodate this, I've replaced the 'project_id' argument with
'project_ids', enabling us to filter multiple projects.

Currently, there is a workaround for this issue. But it's inefficient,
since it needs to fetch billing records for each project separately. We
also need to wrap it with a DB transaction if any of them crash while
generating invoice.

    DB.transaction do
        invoices = project_ids.filter_map do |project_id|
            InvoiceGenerator.new(begin_date, end_date, project_id: project_id).run.first
        end
    end

The new way:

    invoices = InvoiceGenerator.new(begin_date, end_date, project_ids: project_ids).run