Improve invoice charging function
We manually charge invoices. I've resolved some issues we encountered
during the charging process.

Firstly, I switched the logging method from 'puts' to 'Clog'.

We attempted to charge each payment method listed in the billing
information sequentially. Sometimes, the Stripe Ruby SDK throws an error
when a payment fails. In such cases, we need to catch the error and
proceed to charge the next payment method.

If we successfully charge an invoice, there's no need to continue trying
other payment methods, so we break the loop.

If all payment methods fail, we leave the invoice unpaid.