Switch from OpenStruct to double in stripe tests
Exposed by gem upgrades, OpenStruct is not required directly by these
tests (something else must have removed its `require "ostruct"`
expression) and crashed.  I tried to use instance_double per normal,
but somehow Stripe has written their classes such that rather
pedestrian fields are not defined at load time, making RSpec's
`instance_double` unhappy.

I trust this code has been used before and so it is not in error, and
made this unusual use of OpenStruct for this reason to begin with.

To get around this, I use an old fashioned RSpec "double," which
RuboCop reasonably complains about, because the stripe gem's classes
are alone in doing something unusual enough to make `instance_double`
confused.

So, I bloat the code with comment pragmas enabling and disabling that
rule.  As it is four sites, it is sustainable for now, though I
thought about requiring OpenStruct into the tests to keep it the way
it was.