Decrement checkup semaphore if health check succeeds
If checkup semaphore is incremented, we perform a detailed health check on the
resource and go to unavailable state if the health check fails. If the health
check succeeds, we should decrement the checkup semaphore, so that we wouldn't
perform more health checks until the next time the resource is marked for a
detailed health check.