Create Firewalls and attach to Subnet instead of VM
This commit implements the Firewalls move from VMs to Subnets.
Therefore, there are multiple changes regarding model relationships,
Vm::Nexus.assemble, Vnet::SubnetNexus.assemble and finally at the
routes. The changes are not very interesting as they mostly involve
semaphore increments being performed on subnets instead of individual
VMs or entity creations referring to subnets intead of VMs.

One additional small but interesting change is the cidr validation. It
involves 2 changes;
1. Validate IPv6 as well
2. Return the parsed cidr and use its string representation while
creating the record. This is necessary because when NetAddr is able to
parse a cidr like "1.1.1.1/8" without an issue, db insert fails because
the valid form of that cidr is actually "1.0.0.0/8". This used to cause
500 error in console.