Move base_url to config
Current way of generating base_url depends on the request object. However, we
need base_url in contexts where request object is not available, such as while
sending e-mails from background jobs. With this commit, we make base_url a
config variable. Its default value as http://localhost:9292, which is what we
use for development environment and for production environment we can set it to
https://console.ubicloud.com.

With this change, we lose the ability to dynamically set the url scheme, but
we redirect http to https in production anyway.