Add usage alerting based on the total consumption
This commit adds feature of sending alerts when the total consumption exceeds
a user-defined threshold. We associate each alert to a user who created it and
send the email to that user when the alert is triggered. An alternative would
be allowing users to specify an email address to send the alert to, but that
opens an attack vector where a user could spam another user with alerts. This
vector can be mitigated by requiring the user to confirm the email address, but
implementing a whole confirmation workflow would significantly increase the
effort required to implement this feature.