Branch the route with /id instead of /ubid while accessing resources statically
While accessing resources statically (by using their ubid instead of
name) we where branching the URI with /ubid, though thinking about it
further branching with /id looks more correct way to do that.

Users shouldn't care about the format of the unique identifier is in the
ubid format. Using /ubid instead of /id adds an unnecessary confusion
and also doesn't align with the resource scheme we are returning back
from the API.