Replace next_cursor with next_value in the pagination response
Before updating ubid based cursors to polymorphic start_with values, next_cursor
was always in the ubid format. Though naming it as `cursor` feels like it will
be used as a pointer to a resource. So, updating it with next_value to specify
that value will be directly compared while finding the next page's resources.