Replace ubid cursor with polymorphic start_after in pagination request
Client was passing ubid of next page's first resource as cursor to the
paginated list endpoint. Value to be compared was obtained by accessing the
resource with given ubid and getting the value of order column. Concurrent
deletions were broking that pagination as such resources could not be accessed.

In order to fix that issue, cursor, which was expected to be in the ubid
format, has been replaced with polymorphic start_after. Since the start_after will
be directly compared instead of being a pointer to a resource, concurrent deletion
won't cause any issue.

Value of the start_after is passed as a string though the format of it depends on
the order column. If the order column is id, format of start_after will be ubid,
if it is name it will be free text.

For now only id and name (whenever the resource has unique name) columns can
be used as order column, as using non-unique columns doesn't guarantee right
ordering and having disjoint pages.