Remove next_cursor from the pagination response
Response of the pagination has been simplified by removing the next_cursor
from it. Client can pass the last element of the current page as the start_after
of the subsequent request to get the next page. If the items of the page doesn't
have any element, it means that there is no more page and resource.