Allow forcing a specific VmHost to be used for a Vm.
We sometimes want to allocate a Vm on a specific VmHost. This change
allows that.

To do that, we can do:

```
> Prog::Vm::Nexus.assemble(..., force_host_id: vm_host.id)
```

This will force the allocation even if the host is "draining" mode. This
is intentional, since in our common use-case we put the target host in
draining mode to prevent other VMs from being scheduled on it.