Unify VmSetup related executables in rhizome
Merges `prepvm.rb`, 'deletevm.rb`, and `recreate-unpersisted` into one
executable.

These are in preparation to add a 4th action to `setup-vm` which will
reinstall systemd units after an SPDK used for a VM is updated.

We will remove `recreate-unpersisted`, `deletevm.rb`, and `prepvm.rb` in
another commit to allow a graceful upgrade path.