Move setup_environment commands to image generation
The `setup_environment` step involves commands to convert a standard VM
into a runner. Over time, the number of commands has increased in
response to various customer issues, which has also lengthened the run
time.

We previously tried this, but certain commands had to run at runtime
because the runner user was created by cloud-init. Recently, we
separated the runner and runneradmin users. Now, cloud-init creates the
runneradmin user, and we create the runner user in setup_environment.
This allows us to move runner user related commands to image generation.

If the latest image is used, we only run the short version of the
setup_environment script. If an older image is used, we run the long
version of the `setup_environment` script. It allows us to deploy the
new image to some hosts and test it before deploying it to all hosts.

The new version saves the ".setup_info" file using "tee" instead of the
">" redirect, as the file should have different ownership.