Increase nap time from 0 to 1 in donate
With a nap time of 0 in donate, a parent strand executes its children in
a loop until they exit, without taking a nap. This supports children
making progress to unblock the parent.

However, an issue with this is that children are not always able to
immediately make progress and nap 0 becomes wasteful. In production,
we see instances where child strands are executed in a tight loop,
hogging resources and starving other strands.

This change increases nap time in donate to 1. That way, a
parent needs to take a short nap after running its children
and cannot immediately re-execute them.