Do not fail if the runner does not have a strand in serializer
Usually, we don't expect that the runner model will lack a strand.
However, due to eager loading sending multiple queries, there may be
brief periods when the runner doesn't have a strand in the loaded list.
This PR ensures that the serializer doesn't fail if the runner is
without a strand.

With recent changes, this state isn't only the state of virtual machine.
So I moved to the runner model as "display_state" similar to the other
models.

I also rename variables to make the test more
readable.