Use minor version instead of patch version for NodeJS
Every week, a new patch released for NodeJS, then Dependabot creates a
PR to update the version.

Since we are not heavily dependent on NodeJS, we can pin the minor
version instead of the patch version. So, we can avoid the weekly PRs.

While building image, it will pull the latest patch version of the
minor version.