StorageDevice and VmStorageVolume schema updates
We have 2 schema updates here;
1. Make VmStorageVolume.storage_device_id column not nullable;
We occasionally see a drift of StorageDevice.available_storage_gib vs
vm.vm_storage_volume used sizes. This is happening because of bugs and
not maintaining the relationship in between volume and the device
properly. In result, they cause negative available_storage_gib. This
commit introduces new constraints to block these cases so that we can be
aware of the situation.

2. Introduce non zero and smaller than total_storage_gib constraints to
the StorageDevice.available_storage_gib column. Because it makes sense,
just an extra precaution at the DB level.