Replace VmHost.available_storage_gib and total_storage_gib columns with funcs
The available_storage_gib column in the VmHost comes from old times and
have no use in the codebase. It causes confusion and this is cleaning up
that column. Still for the sake of reaching the available storage size
in the host, I added a new function which is a wrapper around
storage_devices.

The total_storage_gib is not used anywhere, so, we are dropping the
column. For the sake of operator tooling, we add a wrapper function in
the host model to find out the total_storage_gib from the storage
devices.