Replace wait_sshable with a script in host
With this commit, we are replacing the logic in wait_sshable in the
controlplane with a script that runs in the host. We punch a small hole
in the firewall for an ip address that is allocated for our use but
specific to the VM. Therefore, this does not create a security issue as
well. This way, we verify the VM connectability without depending on the
controlplane IP address and heroku's networking which does not
support IPv6 (sigh).