Remove Project.provider
We are removing provider property from project. This way, a given
project can have resources in different provider regions. This change
makes sense since we are renaming the location names to not include the
provider information. It will also be helpful for us to run the github
actions workload on hosts from different providers.