Track hugepages used for SPDK installations in the database.
We use number of hugepages used when installing & removing an SPDK
installation. Although this is mostly static, but it is possible that we
decide to change number of hugepages used when installing a new SPDK in
future.

In which case, we should remember what value used for older
installations to update VmHost records correctly when removing them.

This mismatch already happened in production last week, so making this
change to avoid it in future.