Don't increment checkup semaphore if it is already incremented
We had few resources that stays unavailable for a long time, and the checkup
semaphore was incremented every time the pulse_check was called. This was
causing the semaphore to be incremented to a very high value, and the lookups
to semaphore table were taking a long time to finish.