Update error response title with PascalCase type
In the RFC https://datatracker.ietf.org/doc/html/rfc7807, type is
the only key required as a part of the error response. It is used
to decide the kind of the error obtained from the API. Client codes
will be using that type to understand the kind of the error. Note that,
same error code (which is basically the http error code) might have
different error type. To show that is a constant text returned from an
API, value of the type key converted to the PascalCase.

As a side effect of that change, texts shown as title on UI also became
PascalCase. As it still informs the user on the UI as it was informing
before that commit, it is acceptable.