Optimize wait_sshable and firewall rule dance
In this commit, we make a couple of changes to optimize the wait-sshable
and firewall rule application logic at the provisioning time.
1. Immediately run the vm and verify ssh connectivity before firewall
rules are applied.
2. Use push to update firewall rules instead of bud and dummy wait.
3. For the first time we get into wait_sshable, we immediately nap for
15 seconds. This is done with the assumption of wait-sshable will never
take less than 15 seconds and we can just sleep instead of using
compute. It should optimize the amount of compute usage especially at
the busy time.

With the help of these three changes, we are optimizing the amount of time
spent on firewall rule application and the sshability verification.