Optimize nap times while the runner waiting for a vm
Recently, Furkan modified the workflow to update the firewall rule
during VM provisioning. Now, it's updated after the VM reaches the
`wait` label. Consequently, if the runner doesn't check the vm in the
time between the VM reaching the `wait` label and the start of the
firewall rule update, it must wait for the firewall rule update to
complete.

The VM model has a `provisioned_at` timestamp, which provides a better
way to verify whether the VM has been provisioned.

Furthermore, if we nap for 5 seconds at this point, we could potentially
wait an extra 5 seconds, even if the VM is already provisioned, until
the next schedule. We know that our fastest VM provisioning time is 18
seconds. Therefore, if we run this label for the first time, we can wait
for 18 seconds. This PR increases the checking frequency to 1 second
after the initial run, allowing the runner to detect a provisioned VM
more promptly.