Do not use the same runner when the script failed
Currently, when the script failed, we assume it failed because of an
initialization error, and we try to register the same runner again.
This is not always true. The script might be "failed" while running the
workflow. We should create a new spare runner and destroy the failed
one.

I can implement additional checks such as verify if the runner has
completed the job, and avoid creating an spare runner if the job is
completed. However, runner script failures are uncommon. Even when they
occur, they exit with a zero exit code, not a non-zero one. Therefore, I
believe it's currently unnecessary to add more checks. If script
failures increase, we can reconsider adding them.