Download custom images with fixed version
When we download a custom image without specifying a version, defaults
to the latest one, making it difficult to identify which version is in
use on the host.

Moreover, when we release a new version of an image, we don't
immediately want to download the latest version. Instead, we prefer to
test the new version before deploying it on the some hosts first.

Using a fixed version also simplifies the verification of SHA256
checksums.