Refactor VM Page to accommodate the Firewalls in a better way
This commit removes the Network Interfaces box and refactors the
Firewall Rules in the Vm Show page. The main reason we remove the Nics
is that we do not have multiple Nics infrastructure just yet. It is
staying in the main VM page unnecessarily. Also, with the introduction
of the new multiple Firewalls functionality, the main VM page would get
longer. Instead;
1. Since every VM will have a single Private Subnet and the Firewalls
are attached to it, I have moved the private IPv4/6 addresses to the
main details card.
2. Removed the Nics box.
3. Added a new line to see the private subnet.
4. Added a new box for Firewalls to list the firewall and the applied
rules.
5. Remove Firewall rule editing from the VM. This is important because
now firewall rules are applied at the subnet level and editing it
directly in the VM page may result in a change that impacts more than
one resources.