Fix incorrect state transition in PostgresResourceNexus
We used to unconditionally transition to the initialize_certificates label from
refresh_dns_record label, which further transitions to create_billing_record
label. That means every time we refresh the DNS record, we would create a new
root certificates and also new billing records. Both needs to be done once at
the time of the initial provisioning. This commit fixes the issue by adding a
new transition from refresh_dns_record to wait if the initial provisioning is
already done.