Provision standbys in the primary's private subnet
Previously, we created separate subnets for the primary and standby and used
public network for all communication between them. This change provisions the
standby node in the primary's private subnet, which is more secure. It also
enables the standby to use the primary's private IP address for replication,
thus circumventing the user firewalls.