Remove unnecessary detach logic from Nic
Detach nic from the VM logic is not needed since we do not support
managing nics, yet. There is no need to keep it.

I have also moved the cleanup of tunnels to the
association_dependencies.

Furthermore, since we do not bud any child processes anymore, we do not
need to destroy child progs separately at the time of nic destroy.