Delete pages on resolution
We delete(actually destroy) almost all resources when we are done with them. At
some point we considered soft-deleting. As a result of that we didn't deleted
pages on resolution and instead marked them as resolved. However, since then we
also introduced the DeletedRecords concept for historical bookkeeping. I think
keeping the pages around is not worth much anymore. We can always look at the
DeletedRecords if we need to know what was resolved. We also have old pages in
Pagerduty and Slack for easy search. Hence, I'm changing page resolution logic
to delete the page instead of just marking it as resolved.