Fix display state of PostgreSQL resources
We used to just check the label of PostgresResource's own strand. However, the
state of representative server is also important, especially at the time of it
being unavailable. This commit fixes the display state of PostgreSQL resources
by checking the state of the representative server as well.

While doing so, I found another case of wrong display state, while refreshing
the dns records, which is also fixed in this commit.