Don't exit if the pulse check is stuck
Normally, if the pulse check is stuck, the monitor will exit. This is to ensure
that the pulse checking is not stuck indefinitely. It is similar to apoptosis
logic in respirate. However, it happens more frequently than expected and thus
causing stability problems. While investigating the issue, it would be better
to not exit the monitor if the pulse check is stuck. Instead, we will just log
a message and manually cleanup stuck processes for a while.